feat: scoped client chunk related functions to chunk manager
